Do stuff to stop the dependabot madness (#27)
So we do not want to have dependabot working so hard at upgrading
packages in github actions and creating tons of merge conflicts.

Luckily, the Hypermodern author had some fixes for this. Unluckily,
they are not included in the last stable version of the cookiecutter
template, so we didn't get them

Luckily, I have added them by hand here. They are from PR 744 and 746
of
https://github.com/cjolowicz/cookiecutter-hypermodern-python-instance

These do two things:

1. Change all the version requirements "^X.Y.Z" to ">=X.Y.Z" in
pyproject.toml, so as to allow upgrades of major versions
2. Add "versioning-strategy: lockfile-only" to .github/dependabot.yml

I also had to add in a few upper limits of python versions for certain
packages (e.g., cmasher), so that poetry could satisfy all the dependencies